[Refactor] Move IsaacLab PPO helpers to utils
